[AS7-1981] WriteAttributeHandlers for server-group's profile and server-config's group attributes. Affected servers are put in the restart-required state
