+++ host-controller/src/main/java/org/jboss/as/host/controller/HostControllerMessages.java
@@ -25,8 +25,10 @@ package org.jboss.as.host.controller;
 import java.io.File;
 import java.io.IOException;
 import java.net.URL;
+
 import javax.security.sasl.SaslException;
 
+import org.jboss.as.controller.OperationFailedException;
 import org.jboss.as.controller.RunningMode;
 import org.jboss.as.controller.client.helpers.domain.ServerStatus;
 import org.jboss.as.protocol.mgmt.RequestProcessingException;
@@ -534,5 +536,15 @@ public interface HostControllerMessages {
      */
     @Message(id = 10986, value = "Host-Controller is already shutdown.")
     IllegalStateException hostAlreadyShutdown();
+
+    /**
+     * Creates an exception indicating no server group could be found with the given name
+     *
+     * @param groupName the profile name
+     *
+     * @return an {@link OperationFailedException} for the error.
+     */
+    @Message(id = 10987, value = "No server group called: %s")
+    OperationFailedException noServerGroupCalled(String groupName);
 
 }
